+++ src/main/java/com/aig/page/InventoryPage.java
@@ -1,6 +1,5 @@
-package com.aig;
+package com.aig.page;
 
-import com.aig.page.BasePage;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
